SIDE-1195: RG|MM - refactoring to make it working with latest merge from main
