+++ ldk/javascript/examples/consolidate-loop/package.json
@@ -19,7 +19,9 @@
     }
   },
   "dependencies": {
-    "@oliveai/ldk": "file:../.."
+    "@oliveai/ldk": "file:../..",
+    "eslint": "^7.27.0",
+    "eslint-config-prettier": "^8.3.0"
   },
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.13.0",
